+++ Cargo.toml
@@ -6,24 +6,18 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [[bin]]
-name = "start_client"
+name = "client"
 path = "src/start_client.rs"
 
-[[bin]]
-name = "dumb_client"
-path = "src/dumb_client.rs"
 
 [[bin]]
-name = "dumb_client_udp"
-path = "src/dumb_client_udp.rs"
-
-[[bin]]
-name = "start_server"
+name = "server"
 path = "src/start_server.rs"
 
 
 [dependencies]
 bincode = "1.3.3"
+chrono = "0.4.31"
 crossbeam = { version = "0.8.2", features = ["crossbeam-queue"] }
 glam = {version="0.24.2", features=["serde"]}
 lazy_static = "1.4.0"
