+++ src/client/mod.rs
@@ -2,5 +2,7 @@ pub mod draw;
 pub mod event_processing;
 pub mod game;
 pub mod graphics;
+pub mod message_processing;
+pub mod settings;
 pub mod state;
 pub mod udp_networking;
