+++ src/client/settings.rs
@@ -1,0 +1,2 @@
+pub const SERVER_ADDR: &str = "127.0.0.1:8080";
+// const SERVER_ADDR: &str = "72.234.70.195:8081";
