+++ src/common/mod.rs
@@ -1,3 +1,4 @@
 pub mod client_to_server;
 pub mod game_objects;
 pub mod server_to_client;
+pub mod util;
