+++ src/common/util.rs
@@ -1,0 +1,5 @@
+use chrono::Utc;
+
+pub fn get_utc_now() -> i64 {
+    Utc::now().timestamp_millis()
+}
