+++ src/server/message_processing.rs
@@ -1,0 +1,123 @@
+use glam::Vec2;
+
+use crate::{
+    common::{
+        client_to_server::ClientToServerMessageData, game_objects::Player,
+        server_to_client::ServerToClientMessage,
+    },
+    server::enque_outbound_messages::{
+        broadcast_to_all, broadcast_to_all_except, send_to_one_client,
+    },
+};
+
+use super::{state::State, udp_networking::INCOMING_MESSAGE_QUEUE};
+
+pub const DEBUG_PRINT_PROCESSED_MESSAGES: bool = false;
+
+pub async fn process_message_queue(state: &mut State) {
+    // prune_latest_only_messages().await;
+
+    while let Some(message_bundle) = INCOMING_MESSAGE_QUEUE.pop() {
+        let client_id = message_bundle.client_id;
+        match message_bundle.message {
+            ClientToServerMessageData::Connect => {
+                println!("Client {} connected", client_id);
+
+                // send welcome
+                let outbound_message = ServerToClientMessage::Welcome {
+                    server_message: "welcome to the server".to_string(),
+                };
+                send_to_one_client(client_id, outbound_message).await;
+
+                // announce the join
+                let outbound_message = ServerToClientMessage::ClientJoined { id: client_id };
+                broadcast_to_all_except(client_id, outbound_message).await;
+            }
+            ClientToServerMessageData::Disconnect => {
+                println!("Client {} disconnected", client_id);
+
+                // announce the leave
+                let outbound_message = ServerToClientMessage::ClientLeft { id: client_id };
+                broadcast_to_all_except(client_id, outbound_message).await;
+            }
+            ClientToServerMessageData::ChatMessage { message } => {
+                println!("{} says: {}", client_id, message);
+
+                // broadcast the message
+                let outbound_message = ServerToClientMessage::ChatMessage {
+                    from: client_id,
+                    message,
+                };
+                broadcast_to_all_except(client_id, outbound_message).await;
+            }
+            ClientToServerMessageData::RequestToSpawnPlayer => {
+                println!("{} requested to spawn a player", client_id);
+
+                let eid = state.next_eid;
+                state.next_eid += 1;
+
+                // spawn the player
+                let player = Player::new(client_id, eid);
+                state.players.insert(eid, player);
+                println!("spawned player {}", eid);
+
+                // announce the spawn
+                let outbound_message = ServerToClientMessage::SpawnPlayer {
+                    owner_client_id: client_id,
+                    entity_id: eid,
+                    pos: Vec2::ZERO,
+                };
+                broadcast_to_all(outbound_message).await;
+            }
+            ClientToServerMessageData::EntityPosition { entity_id, pos } => {
+                if let Some(player) = state.players.get_mut(&entity_id) {
+                    player.pos = pos;
+                }
+
+                let outbound_message = ServerToClientMessage::EntityPosition { entity_id, pos };
+                broadcast_to_all_except(client_id, outbound_message).await;
+            }
+            ClientToServerMessageData::RequestAllPlayers => {
+                println!("{} requested all players", client_id);
+
+                let players = state.players.values().cloned().collect();
+
+                let outbound_message = ServerToClientMessage::AllPlayers { players };
+                send_to_one_client(client_id, outbound_message).await;
+            }
+        }
+    }
+}
+
+// pub async fn prune_latest_only_messages() {
+//     let queue = INCOMING_MESSAGE_QUEUE.clone();
+
+//     let mut num_pruned = 0;
+//     let mut latest_messages: HashMap<u8, ClientToServerMessage> = HashMap::new();
+//     let mut to_retain: Vec<ClientToServerMessage> = Vec::new();
+
+//     while let Some(bundle) = queue.pop() {
+//         if let Some(keep_latest_only_id) = get_keep_latest_only_message_type_id(&bundle.message) {
+//             latest_messages.insert(keep_latest_only_id, bundle.clone());
+//             num_pruned += 1;
+//         } else {
+//             to_retain.push(bundle);
+//         }
+//     }
+//     // println!("num pruned {}", num_pruned);
+
+//     // Add the latest messages back to the retained list
+//     to_retain.extend(latest_messages.values().cloned());
+
+//     // Refill the queue with the pruned list
+//     for bundle in to_retain.iter() {
+//         queue.push(bundle.clone()).unwrap();
+//     }
+// }
+
+// pub fn get_keep_latest_only_message_type_id(message: &ClientToServerMessageData) -> Option<u8> {
+//     match message {
+//         ClientToServerMessageData::EntityPosition { .. } => Some(0),
+//         _ => None,
+//     }
+// }
