+++ src/server/mod.rs
@@ -1,6 +1,7 @@
 pub mod client_bookkeeping;
 pub mod enque_outbound_messages;
 pub mod game;
+pub mod message_processing;
 pub mod settings;
 pub mod state;
 pub mod udp_networking;
